Move test to behaviour file
